bug fix attempt 2

+++ event_producer/cmsconnect_miniaods_UL17/fragments/BulkGravitonToHHTo4W_MX-Var_MH-260to650_JHUVariableWMass.py
@@ -32,8 +32,8 @@ num_low_points = len(low_m_higgs)
 num_high_points = len(m_higgs)
 
 # solve system of equations s.t. 1) total weight sums to 1, and 2) the first weight is 1 / (# of low points) i.e. same # of events as 260 GeV
-m = np.array([[num_high_points, num_high_points * (num_high_points - 1) / 2], [1, num_high_points]])
-b = np.array([1, 1 / num_low_points])
+m = np.array([[num_high_points, num_high_points * (num_high_points - 1) / 2], [1, num_high_points]]).astype(float)
+b = np.array([1, 1 / num_low_points]).astype(float)
 # a is smallest weight, d is spacing between weights
 a, d = np.linalg.solve(m, b)
